+++ scio-examples/src/main/scala/com/spotify/scio/examples/extra/WordCountScioIO.scala
@@ -57,7 +57,7 @@ object WordCountScioIO {
       .map { w =>
         // Trim input lines, update distribution metric
         val trimmed = w.trim
-        lineDist.update(trimmed.length)
+        lineDist.update(trimmed.length.toLong)
         trimmed
       }
       .filter { w =>
